+++ crates/holochain/src/conductor/interface/websocket.rs
@@ -1229,7 +1229,7 @@ mod test {
                     }
                 }
             })
-            .last()
+            .next_back()
             .unwrap()
     }
 }